+++ php/extras/sqlite.php
@@ -719,8 +719,14 @@ function sqliteEnumQueryResults($data,$params=array()){
 	//write to file or return a recordset?
 	if(isset($params['-filename'])){
 		$starttime=microtime(true);
-		if(file_exists($params['-filename'])){unlink($params['-filename']);}
-    	$fh = fopen($params['-filename'],"wb");
+		if(isset($params['-append'])){
+			//append
+    		$fh = fopen($params['-filename'],"ab");
+		}
+		else{
+			if(file_exists($params['-filename'])){unlink($params['-filename']);}
+    		$fh = fopen($params['-filename'],"wb");
+		}
     	if(!isset($fh) || !is_resource($fh)){
 			odbc_free_result($result);
 			return 'hanaQueryResults error: Failed to open '.$params['-filename'];
